Refactor settings module to split the settings routes into different files
